Fork: changing swap functionality for arm64
On 64-bit arm, register `x18` is a platform register, reserved for the
use of platform ABI. On aarch64 specifically, the apple documentation
also calls out `x18` as a reserved platform register and explicitly says
"Don't use this register". Currently for arm64 assembly codegen, a swap
operation was implemented as 3 `mov`s, using `x18` as a temporary
register. This can possibly lead to some UB at runtime as the contents
of `x18` may not be as expected.

Rather than performing a swap using 3 `mov`s and a temp, instead
implement it using `stp`/`ldp` to leverage the stack. It is admittedly
less performant than using a temporary register but is worth it compared
to potential UB. An alternate approach would be to use a register other
than hardcoding `x18`, which would require additional logic at
register-allocation time.